Seccion 6 Authorization and Authentication

+++ src/tasks/tasks.service.ts
@@ -6,6 +6,7 @@ import { InjectRepository } from '@nestjs/typeorm';
 import { Task } from './task.entity';
 import { TaskStatus } from './task-status.enum';
 import { GetTasksFilterDto } from './dto/get-tasks-filter.dto';
+import { User } from 'src/auth/user.entity';
 
 @Injectable()
 export class TasksService {
@@ -16,8 +17,11 @@ export class TasksService {
     //private tasks: Task[] = [];
 
 
-    async getTasks(filterDto: GetTasksFilterDto): Promise<Task[]>{
-      return this.taskRepository.getTasks(filterDto);    
+    async getTasks(
+        filterDto: GetTasksFilterDto,
+        user: User,
+        ): Promise<Task[]>{
+      return this.taskRepository.getTasks(filterDto, user);    
     }
 
  /*
@@ -35,32 +39,41 @@ export class TasksService {
         }
         return tasks;
     }*/
-    async getTaskById(id: number): Promise<Task>{
-        const found = await this.taskRepository.findOne(id);
+    async getTaskById(id: number, user: User): Promise<Task>{
+        const found = await this.taskRepository.findOne({ where: { id, userId: user.id }});
         if(!found){
             throw new NotFoundException(`Task with ID "${id}" not found`);
         }
         return found;
     }
 
-    async createTask(createTaskDto: CreateTaskDto): Promise<Task>{
-        return this.taskRepository.createTask(createTaskDto);
+    async createTask(
+        createTaskDto: CreateTaskDto,
+        user: User,   
+        ): Promise<Task>{
+        return this.taskRepository.createTask(createTaskDto, user);
     }
-    async deleteTask(id: number): Promise<void>{
-        const result = await this.taskRepository.delete(id);
+    async deleteTask(id: number,
+        user: User,
+        ): Promise<void>{
+        const result = await this.taskRepository.delete({ id, userId: user.id});
         if(result.affected === 0){
             throw new NotFoundException(`Task with ID "${id}" not found`);
         }
         console.log(result);
     }
 
-    async updateTaskStatus(id: number, status: TaskStatus): Promise<Task>{
-        const task= await this.getTaskById(id);
+     async updateTaskStatus(
+         id: number, 
+         status: TaskStatus,
+         user: User,
+        ): Promise<Task>{
+        const task= await this.getTaskById(id, user);
         task.status = status;
         await task.save();
         return task;
     }
-
+ 
     
     /*createTask(createTaskDto: CreateTaskDto): Task{
         const { title, description } = createTaskDto;
